requirements.txt
langchain-core
langgraph
faiss-cpu
sentence-transformers
python-dotenv
numpy
pydantic
google-genai